+++ src/js/utils.js
@@ -28,9 +28,9 @@ function parseFunctionFromExpression( expr, func ) {
         openBrackets = 1;
         args = [];
         start = result.index;
-        index = findFunc.lastIndex;
-        argStart = index;
-        while ( openBrackets !== 0 ) {
+        argStart = findFunc.lastIndex;
+        index = argStart - 1;
+        while ( openBrackets !== 0 && index < expr.length ) {
             index++;
             if ( expr[ index ] === '(' ) {
                 openBrackets++;
@@ -42,7 +42,9 @@ function parseFunctionFromExpression( expr, func ) {
             }
         }
         // add last argument
-        args.push( expr.substring( argStart, index ).trim() );
+        if ( argStart < index ) {
+            args.push( expr.substring( argStart, index ).trim() );
+        }
 
         // add [ 'function( a ,b)', ['a','b'] ] to result array
         results.push( [ expr.substring( start, index + 1 ), args ] );